refactor: ToiletAccessibility → Toilet API 전환 + PR #170 리뷰 반영
- scc-api submodule 업데이트(d8c0919) + codegen: searchToilets/getToilet,
  ToiletSummaryDto/ToiletDetailDto, ToiletApi 사용
- ExternalAccessibilityDetailScreen → ToiletDetailScreen 으로 rename
  (디렉토리/컴포넌트/네비게이션 등록/라우트 파라미터 toiletId)
- sourceType 분기 제거, 섹션별 데이터 존재 여부로 조건부 렌더
  (이미지/locationComment/comment/공공상세 toiletDetails)
- getToilet 로딩/에러(ToastUtils.showOnApiError + 폴백 UI) 처리
- ToiletCard: sourceType 분기 제거, available 데이터 유무로 조건부 렌더
- useSearchRequest: searchToilets 호출, 위치 가드(없으면 toast 후 [] 반환),
  non-null 단언 제거, 불필요 주석 제거
- ToiletSection: 위치/기타 참고사항 라벨 문구 반영
- AppComponentsContext/App.tsx: toiletApi 로 주입명 정리
