+++ src/screens/ToiletDetailScreen/AvailableLabel.tsx
