+++ src/screens/ToiletMapScreen/data.ts
@@ -1,14 +1,12 @@
 import {MarkerItem} from '@/components/maps/MarkerItem.ts';
 import {
   ExternalAccessibility,
-  ToiletAccessibilityDetailDto,
-  ToiletAccessibilitySourceTypeDto,
-  ToiletAccessibilitySummaryDto,
+  ToiletAccessibilityDetails,
+  ToiletSummaryDto,
 } from '@/generated-sources/openapi';
 
 export type ToiletDetails = Omit<
   ExternalAccessibility & {
-    sourceType: ToiletAccessibilitySourceTypeDto;
     imageUrl?: string;
     gender?: {
       state: 'MALE' | 'FEMALE' | 'BOTH';
@@ -128,7 +126,6 @@ export function mapToToiletDetails(
       : undefined;
   return {
     ...toilet,
-    sourceType: ToiletAccessibilitySourceTypeDto.ExternalAccessibility,
     imageUrl: toilet.toilet_details?.image_url,
     available,
     gender,
@@ -154,38 +151,40 @@ export function mapToToiletDetails(
 }
 
 /**
- * 통합 화장실 상세(`/getToiletAccessibility`) 응답(`ToiletAccessibilityDetailDto`, EXTERNAL 소스)을
- * 기존 TDP 렌더에 쓰이는 `ToiletDetails`로 매핑한다.
+ * 화장실 상세(`/getToilet`)의 공공데이터 상세(`toiletDetails`, `ToiletAccessibilityDetails`)를
+ * TDP 의 "화장실 사용/접근/내부 정보" 섹션 렌더에 쓰이는 `ToiletDetails`로 매핑한다.
  */
-export function mapDetailToToiletDetails(
-  detail: ToiletAccessibilityDetailDto,
+export function mapToiletDetailsToToiletDetails(
+  id: string,
+  name: string,
+  address: string | undefined,
+  location: ExternalAccessibility['location'],
+  toiletDetails: ToiletAccessibilityDetails,
 ): ToiletDetails {
   const externalShaped: ExternalAccessibility = {
-    id: detail.id,
-    name: detail.name,
-    address: detail.address ?? undefined,
-    location: detail.location,
+    id,
+    name,
+    address,
+    location,
     category: 'TOILET',
-    toilet_details: detail.toiletDetails,
+    toilet_details: toiletDetails,
   };
   return mapToToiletDetails(externalShaped);
 }
 
 /**
- * 통합 화장실 검색(`/searchToiletAccessibilities`) 결과(`ToiletAccessibilitySummaryDto`)를
+ * 화장실 검색(`/searchToilets`) 결과(`ToiletSummaryDto`)를
  * 카드/마커 렌더에 쓰이는 `ToiletDetails & MarkerItem`으로 매핑한다.
- * USER 소스는 성별/입구 등 상세 필드가 없으므로 태그/사용가능 라벨 없이 이름/주소/썸네일만 채운다.
+ * 검색 응답에는 성별/입구 등 상세 필드가 없으므로 이름/주소만 채운다.
  */
 export function mapSummaryToToiletDetails(
-  summary: ToiletAccessibilitySummaryDto,
+  summary: ToiletSummaryDto,
 ): ToiletDetails & MarkerItem {
   return {
     id: summary.id,
-    sourceType: summary.sourceType,
     name: summary.name,
     address: summary.address ?? undefined,
     location: summary.location,
-    imageUrl: summary.thumbnailUrl ?? undefined,
     displayName: summary.name,
     markerIcon: {
       icon: 'toilet',
